pyproject.toml
[project]
name = "whisper-finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.3.0",
    "datasets[audio]>=3.2.0",
    "evaluate>=0.4.3",
    "jiwer>=3.1.0",
    "numpy==1.23.5",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
    "tensorboard>=2.18.0",
    "transformers>=4.48.2",
]
